remove mock-orig-dst feature
