+++ linkerd/app/Cargo.toml
@@ -13,7 +13,6 @@ This is used by tests and the executable.
 
 [features]
 allow-loopback = ["linkerd-app-outbound/allow-loopback"]
-mock-orig-dst  = ["linkerd-app-core/mock-orig-dst"]
 
 [dependencies]
 futures = "0.3.9"
