+++ linkerd/app/integration/Cargo.toml
@@ -24,8 +24,8 @@ http = "0.2"
 http-body = "0.4"
 hyper = { version = "0.14.2", features = ["http1", "http2", "stream", "client", "server"] }
 linkerd-channel = { path = "../../channel" }
-linkerd-app = { path = "..", features = ["allow-loopback", "mock-orig-dst"] }
-linkerd-app-core = { path = "../core", features = ["mock-orig-dst"] }
+linkerd-app = { path = "..", features = ["allow-loopback"] }
+linkerd-app-core = { path = "../core" }
 linkerd-metrics = { path = "../../metrics", features = ["test_util"] }
 linkerd2-proxy-api = { git = "https://github.com/linkerd/linkerd2-proxy-api", tag = "v0.1.18", features = ["arbitrary"] }
 linkerd-app-test = { path = "../test" }
