+++ linkerd/app/test/Cargo.toml
@@ -23,7 +23,7 @@ http = "0.2"
 http-body = "0.4"
 hyper = { version = "0.14.2", features = ["http1", "http2"] }
 linkerd-channel = { path = "../../channel" }
-linkerd-app-core = { path = "../core", features = ["mock-orig-dst"] }
+linkerd-app-core = { path = "../core" }
 linkerd-identity = { path = "../../identity" }
 linkerd-io = { path = "../../io", features = ["tokio-test"] }
 regex = "1"
