+++ linkerd2-proxy/Cargo.toml
@@ -9,7 +9,6 @@ description = "The main proxy executable"
 
 [features]
 default = ["multicore"]
-mock-orig-dst  = ["linkerd-app/mock-orig-dst"]
 multicore = ["tokio/rt-multi-thread", "num_cpus"]
 
 [dependencies]
